+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.0.0 2017-10-03
+
+  * `SMTPServerAsPromised` no more extends `SMTPServer`. The `SMTPServer`
+    object is stored in `server` property.
+  * Exports also as a class and namespace and the default.
+  * Typings for Typescript.
+  * Based on promise-readable@1.x.x
+
 ## v0.1.2 2017-06-22
 
   * Upgraded chai-as-promised@7.0.0, promise-readable@0.4.3, tap@10.5.1,